+++ stencil-workspace/src/components.d.ts
@@ -11,6 +11,7 @@ import { App } from "./components/modus-navbar/apps-menu/modus-navbar-apps-menu"
 import { App as App1 } from "./components/modus-navbar/apps-menu/modus-navbar-apps-menu";
 import { RadioButton } from "./components/modus-radio-group/modus-radio-button";
 import { Tab } from "./components/modus-tabs/modus-tabs";
+import { TreeViewItemOptions } from "./components/modus-content-tree/types";
 export namespace Components {
     interface ModusAccordion {
         /**
@@ -720,9 +721,13 @@ export namespace Components {
          */
         "checkboxSelection": boolean;
         /**
-          * (optional) Disables each tree item
+          * (optional) Checked tree items by default
          */
-        "disabled": boolean;
+        "checkedItems": string[];
+        /**
+          * (optional) Expanded tree items by default
+         */
+        "expandedItems": string[];
         /**
           * (optional) Enables multiple checkbox selection
          */
@@ -731,6 +736,10 @@ export namespace Components {
           * (optional) Enables multiple tree items selection
          */
         "multiSelection": boolean;
+        /**
+          * (optional) Selected tree items by default
+         */
+        "selectedItems": string[];
         /**
           * (optional) The default size of all tree items
          */
@@ -749,6 +758,7 @@ export namespace Components {
           * (optional) Expanded state of the tree item
          */
         "expanded": boolean;
+        "focusItem": () => Promise<void>;
         /**
           * (optional) Checkbox indeterminate state of the tree item
          */
@@ -1913,9 +1923,13 @@ declare namespace LocalJSX {
          */
         "checkboxSelection"?: boolean;
         /**
-          * (optional) Disables each tree item
+          * (optional) Checked tree items by default
          */
-        "disabled"?: boolean;
+        "checkedItems"?: string[];
+        /**
+          * (optional) Expanded tree items by default
+         */
+        "expandedItems"?: string[];
         /**
           * (optional) Enables multiple checkbox selection
          */
@@ -1924,6 +1938,10 @@ declare namespace LocalJSX {
           * (optional) Enables multiple tree items selection
          */
         "multiSelection"?: boolean;
+        /**
+          * (optional) Selected tree items by default
+         */
+        "selectedItems"?: string[];
         /**
           * (optional) The default size of all tree items
          */